requirement.txt
python-dotenv
opencv-python
requests
cloudinary
ultralytics
insightface
onnxruntime
faiss-cpu==1.7.4
numpy==1.26.4
flask
filterpy